// src/type/pull-request.type.ts
export type PRType = 'NEW_CHAPTER' | 'EDIT_CHAPTER' | 'DELETE_CHAPTER';

export type PRStatus = 'OPEN' | 'APPROVED' | 'REJECTED' | 'CLOSED' | 'MERGED';

export type PRLabel = 'NEEDS_REVIEW' | 'QUALITY_ISSUE' | 'GRAMMAR' | 'PLOT_HOLE' | 'GOOD_FIRST_PR';